refactor: use internal util
