+++ frontend/cs/r1cs/api.go
@@ -28,9 +28,9 @@ import (
 	"github.com/consensys/gnark/backend"
 	"github.com/consensys/gnark/backend/hint"
 	"github.com/consensys/gnark/frontend"
-	"github.com/consensys/gnark/frontend/utils"
 	"github.com/consensys/gnark/internal/backend/compiled"
 	"github.com/consensys/gnark/internal/parser"
+	"github.com/consensys/gnark/internal/utils"
 )
 
 // Add returns res = i1+i2+...in
