+++ frontend/cs/r1cs/assertions.go
@@ -21,8 +21,8 @@ import (
 	"math/big"
 
 	"github.com/consensys/gnark/frontend"
-	"github.com/consensys/gnark/frontend/utils"
 	"github.com/consensys/gnark/internal/backend/compiled"
+	"github.com/consensys/gnark/internal/utils"
 )
 
 // AssertIsEqual adds an assertion in the constraint system (i1 == i2)
