+++ internal/generator/backend/template/representations/solution.go.tmpl
@@ -9,7 +9,7 @@ import (
 
     "github.com/consensys/gnark/backend/hint"
     "github.com/consensys/gnark/internal/backend/compiled"
-	"github.com/consensys/gnark/frontend/utils"
+	"github.com/consensys/gnark/internal/utils"
     
 	{{ template "import_fr" . }}
 	{{ template "import_curve" . }}
