+++ std/algebra/twistededwards/curve.go
@@ -27,7 +27,7 @@ import (
 	edbn254 "github.com/consensys/gnark-crypto/ecc/bn254/twistededwards"
 	edbw6633 "github.com/consensys/gnark-crypto/ecc/bw6-633/twistededwards"
 	edbw6761 "github.com/consensys/gnark-crypto/ecc/bw6-761/twistededwards"
-	"github.com/consensys/gnark/frontend"
+	"github.com/consensys/gnark/internal/utils"
 )
 
 // EdCurve stores the info on the chosen edwards curve
@@ -67,12 +67,12 @@ func newEdBN254() EdCurve {
 	edcurve.Cofactor.FromMont()
 
 	return EdCurve{
-		A:        frontend.FromInterface(edcurve.A),
-		D:        frontend.FromInterface(edcurve.D),
-		Cofactor: frontend.FromInterface(edcurve.Cofactor),
-		Order:    frontend.FromInterface(edcurve.Order),
-		BaseX:    frontend.FromInterface(edcurve.Base.X),
-		BaseY:    frontend.FromInterface(edcurve.Base.Y),
+		A:        utils.FromInterface(edcurve.A),
+		D:        utils.FromInterface(edcurve.D),
+		Cofactor: utils.FromInterface(edcurve.Cofactor),
+		Order:    utils.FromInterface(edcurve.Order),
+		BaseX:    utils.FromInterface(edcurve.Base.X),
+		BaseY:    utils.FromInterface(edcurve.Base.Y),
 		ID:       ecc.BN254,
 	}
 
@@ -84,12 +84,12 @@ func newEdBLS381() EdCurve {
 	edcurve.Cofactor.FromMont()
 
 	return EdCurve{
-		A:        frontend.FromInterface(edcurve.A),
-		D:        frontend.FromInterface(edcurve.D),
-		Cofactor: frontend.FromInterface(edcurve.Cofactor),
-		Order:    frontend.FromInterface(edcurve.Order),
-		BaseX:    frontend.FromInterface(edcurve.Base.X),
-		BaseY:    frontend.FromInterface(edcurve.Base.Y),
+		A:        utils.FromInterface(edcurve.A),
+		D:        utils.FromInterface(edcurve.D),
+		Cofactor: utils.FromInterface(edcurve.Cofactor),
+		Order:    utils.FromInterface(edcurve.Order),
+		BaseX:    utils.FromInterface(edcurve.Base.X),
+		BaseY:    utils.FromInterface(edcurve.Base.Y),
 		ID:       ecc.BLS12_381,
 	}
 
@@ -101,12 +101,12 @@ func newEdBLS377() EdCurve {
 	edcurve.Cofactor.FromMont()
 
 	return EdCurve{
-		A:        frontend.FromInterface(edcurve.A),
-		D:        frontend.FromInterface(edcurve.D),
-		Cofactor: frontend.FromInterface(edcurve.Cofactor),
-		Order:    frontend.FromInterface(edcurve.Order),
-		BaseX:    frontend.FromInterface(edcurve.Base.X),
-		BaseY:    frontend.FromInterface(edcurve.Base.Y),
+		A:        utils.FromInterface(edcurve.A),
+		D:        utils.FromInterface(edcurve.D),
+		Cofactor: utils.FromInterface(edcurve.Cofactor),
+		Order:    utils.FromInterface(edcurve.Order),
+		BaseX:    utils.FromInterface(edcurve.Base.X),
+		BaseY:    utils.FromInterface(edcurve.Base.Y),
 		ID:       ecc.BLS12_377,
 	}
 
@@ -118,12 +118,12 @@ func newEdBW633() EdCurve {
 	edcurve.Cofactor.FromMont()
 
 	return EdCurve{
-		A:        frontend.FromInterface(edcurve.A),
-		D:        frontend.FromInterface(edcurve.D),
-		Cofactor: frontend.FromInterface(edcurve.Cofactor),
-		Order:    frontend.FromInterface(edcurve.Order),
-		BaseX:    frontend.FromInterface(edcurve.Base.X),
-		BaseY:    frontend.FromInterface(edcurve.Base.Y),
+		A:        utils.FromInterface(edcurve.A),
+		D:        utils.FromInterface(edcurve.D),
+		Cofactor: utils.FromInterface(edcurve.Cofactor),
+		Order:    utils.FromInterface(edcurve.Order),
+		BaseX:    utils.FromInterface(edcurve.Base.X),
+		BaseY:    utils.FromInterface(edcurve.Base.Y),
 		ID:       ecc.BW6_633,
 	}
 
@@ -135,12 +135,12 @@ func newEdBW761() EdCurve {
 	edcurve.Cofactor.FromMont()
 
 	return EdCurve{
-		A:        frontend.FromInterface(edcurve.A),
-		D:        frontend.FromInterface(edcurve.D),
-		Cofactor: frontend.FromInterface(edcurve.Cofactor),
-		Order:    frontend.FromInterface(edcurve.Order),
-		BaseX:    frontend.FromInterface(edcurve.Base.X),
-		BaseY:    frontend.FromInterface(edcurve.Base.Y),
+		A:        utils.FromInterface(edcurve.A),
+		D:        utils.FromInterface(edcurve.D),
+		Cofactor: utils.FromInterface(edcurve.Cofactor),
+		Order:    utils.FromInterface(edcurve.Order),
+		BaseX:    utils.FromInterface(edcurve.Base.X),
+		BaseY:    utils.FromInterface(edcurve.Base.Y),
 		ID:       ecc.BW6_761,
 	}
 
@@ -152,12 +152,12 @@ func newEdBLS315() EdCurve {
 	edcurve.Cofactor.FromMont()
 
 	return EdCurve{
-		A:        frontend.FromInterface(edcurve.A),
-		D:        frontend.FromInterface(edcurve.D),
-		Cofactor: frontend.FromInterface(edcurve.Cofactor),
-		Order:    frontend.FromInterface(edcurve.Order),
-		BaseX:    frontend.FromInterface(edcurve.Base.X),
-		BaseY:    frontend.FromInterface(edcurve.Base.Y),
+		A:        utils.FromInterface(edcurve.A),
+		D:        utils.FromInterface(edcurve.D),
+		Cofactor: utils.FromInterface(edcurve.Cofactor),
+		Order:    utils.FromInterface(edcurve.Order),
+		BaseX:    utils.FromInterface(edcurve.Base.X),
+		BaseY:    utils.FromInterface(edcurve.Base.Y),
 		ID:       ecc.BLS24_315,
 	}
